Move set_parameters to Simulation
