+++ docs/user_guide/remote_access_devices.rst
@@ -52,11 +52,12 @@ command line:
     $ lewis-control device target_speed 100
     $ lewis-control device start
 
-It is possible to set multiple device parameters at once:
+It is possible to set multiple device parameters at once, but this goes through the simulation
+itself, so that it is generic to all devices:
 
 ::
 
-    $ lewis-control device set_parameters "{'target_speed': 1, 'target_phase': 20}"
+    $ lewis-control simulation set_device_parameters "{'target_speed': 1, 'target_phase': 20}"
 
 
 Value Interpretation and Syntax
